Add user not found test and replace int64 to int

+++ store/interface.go
@@ -3,9 +3,9 @@ package store
 import "github.com/ssshekhu53/user-detail-management/models"
 
 type User interface {
-	Create(*models.UserRequest) int64
+	Create(*models.UserRequest) int
 	Get(filters *models.Filters) []models.User
-	GetByID(int64) (*models.User, error)
+	GetByID(int) (*models.User, error)
 	Update(*models.User)
-	Delete(int64)
+	Delete(int)
 }
